+++ ext/stub/interfaces/implementinterface.zep.c
@@ -59,7 +59,7 @@ PHP_METHOD(Stub_Interfaces_ImplementInterface, getVoid)
 		Z_PARAM_OBJECT_OF_CLASS(obj, stub_interfaces_interfaceint_ce)
 	ZEND_PARSE_PARAMETERS_END();
 	zephir_fetch_params_without_memory_grow(1, 0, &obj);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 36, obj);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 38, obj);
 }
 
 PHP_METHOD(Stub_Interfaces_ImplementInterface, getObj)