+++ ext/stub/reflection.zep.c
@@ -78,7 +78,7 @@ PHP_METHOD(Stub_Reflection, setReflectionParameter)
 		Z_PARAM_OBJECT_OF_CLASS(parameter, zephir_get_internal_ce(SL("reflectionparameter")))
 	ZEND_PARSE_PARAMETERS_END();
 	zephir_fetch_params_without_memory_grow(1, 0, &parameter);
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 92, parameter);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 94, parameter);
 	RETURN_MEMBER(getThis(), "reflectionParameter");
 }