+++ ext/stub/traits/counteruser.zep.c
@@ -52,10 +52,10 @@ PHP_METHOD(Stub_Traits_CounterUser, increment)
 	if (UNEXPECTED(!_zephir_prop_0)) {
 		_zephir_prop_0 = zend_string_init("count", 5, 1);
 	}
-	zephir_read_property_cached(&_0, this_ptr, _zephir_prop_0, 128, PH_NOISY_CC | PH_READONLY);
+	zephir_read_property_cached(&_0, this_ptr, _zephir_prop_0, 130, PH_NOISY_CC | PH_READONLY);
 	ZVAL_UNDEF(&_1);
 	ZVAL_LONG(&_1, (zephir_get_numberval(&_0) + 2));
-	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 128, &_1);
+	zephir_update_property_zval_cached(this_ptr, _zephir_prop_0, 130, &_1);
 	RETURN_MEMBER_TYPED(getThis(), "count", IS_LONG);
 }
 
@@ -84,7 +84,7 @@ PHP_METHOD(Stub_Traits_CounterUser, tagged)
 
 	ZEPHIR_CALL_METHOD(&_0, this_ptr, "label", NULL, 0);
 	zephir_check_call_status();
-	zephir_read_property_cached(&_1, this_ptr, _zephir_prop_0, 128, PH_NOISY_CC | PH_READONLY);
+	zephir_read_property_cached(&_1, this_ptr, _zephir_prop_0, 130, PH_NOISY_CC | PH_READONLY);
 	ZEPHIR_CALL_FUNCTION(&_2, "strval", NULL, 61, &_1);
 	zephir_check_call_status();
 	ZEPHIR_CONCAT_VSV(return_value, &_0, ":", &_2);